+++ src/dune_rules/gen_rules.ml
@@ -277,6 +277,7 @@ let gen_rules sctx dir_contents cctxs expander
           Coq_rules.setup_extraction_rules ~sctx ~dir:ctx_dir ~dir_contents m
         | Coq_stanza.Coqpp.T m ->
           Coq_rules.setup_coqpp_rules ~sctx ~dir:ctx_dir m
+        | Coq_stanza.Ffi.T m -> Coq_rules.setup_ffi_rules ~sctx ~dir:ctx_dir m
         | _ -> Memo.return ())
   in
   let+ () = define_all_alias ~dir:ctx_dir ~project ~js_targets in